requirements.txt
grpcio == 1.3.5
grpcio-tools == 1.3.5
protobuf>=3.2.0
leveldb == 0.193
coloredlogs
verboselogs
cryptography == 1.8.1
PyJWT
pyopenssl
yappi
GitPython >= 2.1.1
Flask
Flask-RESTful
requests
fluent-logger >= 0.5.0
gunicorn
setproctitle
python-pkcs11